Add logging for commit selection
